Make AlluxioWorker a runner class
